refacto: renamed pret/ui/* to pret_* (and pret.react for react)
# Conflicts:
#	docs/tutorials/widgets.md
